gildedrosepython/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
django==3.2.8
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
pytz==2021.3
    # via django
sqlparse==0.4.2
    # via django
